% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chea_example}
\alias{chea_example}
\title{ChEA gene set library}
\format{A list of 384 character vectors:
\describe{
  \item{gene set name}{contains gene names associated with the ChIP-seq experiment}
  ...
}}
\source{
\url{http://amp.pharm.mssm.edu/Enrichr/}
}
\usage{
chea_example
}
\description{
May be used for genesetr examples.
}
\details{
A list of gene sets manually curated from transcription factor
ChIP-seq studies in the published literature.
Downloaded from the web tool Enrichr on Aug. 23, 2017.
ChEA_2015.txt
}
\keyword{datasets}
