% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_libs.R
\name{checkLib}
\alias{checkLib}
\title{Check library format compliance with genesetr formats.}
\usage{
checkLib(x)
}
\arguments{
\item{filename}{Library object.}
}
\value{
TRUE if compliant, FALSE with accompanying warning if not compliant.

The first column is the gene members and the second caolumn is the gene set names.
}
\description{
Check library format compliance with genesetr formats.
}
\examples{
gmt = loadLongDF(myLib)
print(checkLib(myLib))
}
