% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{encode_example}
\alias{encode_example}
\title{ENCODE gene set library}
\format{A list of 816 character vectors:
\describe{
  \item{gene set name}{contains gene names associated with the ChIP-seq experiment}
  ...
}}
\source{
\url{http://amp.pharm.mssm.edu/Enrichr/}
}
\usage{
encode_example
}
\description{
May be used for genesetr examples.
}
\details{
A list of gene sets from transcription factor
ChIP-seq studies conducted by the ENCODE consortium.
Downloaded from the web tool Enrichr on Aug. 23, 2017.
ENCODE_TF_ChIP-seq_2015.txt
}
\keyword{datasets}
