% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_gene_names.R
\name{ensembl2HGNC}
\alias{ensembl2HGNC}
\title{Maps Ensembl transcript and gene ids to HGNC-approved gene symbols}
\usage{
ensembl2HGNC(ids, untranslatable.na = F)
}
\arguments{
\item{ids}{a character vector of ensembl ids}

\item{untranslatable.na}{boolean, default is false. If an id does not have an HGNC
mapping, then the function will return an NA for that gene}
}
\value{
a character vector with all ids mapped to approved HGNC symbols
}
\description{
Maps Ensembl transcript and gene ids to HGNC-approved gene symbols
}
\examples{
ids = genesetr::ensembl_ids$ensembl_gene_id[1:10]
HGNC_symbols = ensembl2HGNC(ids, untranslatable.na = T)

}
