% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_tests.R
\name{fastFET}
\alias{fastFET}
\title{Fast version of Fisher Exact Test}
\usage{
fastFET(a, b, c, d, alternative = "greater")
}
\arguments{
\item{a}{A vector of values where a_i corresponds}

\item{b}{A vector of values where b_i corresponds to}

\item{c}{A vector of values where c\_{i} corresponds to}

\item{d}{A vector of values where d_{i} corresponds to}
}
\value{
A vector of p-values P={p_1, p_2,...,p_n-1,p_n} where p_i corresponds to the FET result of c_i.
}
\description{
accomodates a set of n 2x2 contingency tables C = \{c_1,c_2,...,c_n-1,c_n\}, with each c_i of the form:
           setAn ¬setA
       setB  a     b | a+b
      ¬setB  c     d | c+d
         ------------|-----
            a+c   b+d| a+b+c+d
}
\examples{
Given three contingency tables:
     setA  ¬setA          setA  ¬setA         setB  ¬setB
setB  3     100        setC  a     b        setD  a     b
¬setB  123   500       ¬setC  c     d       ¬setD  c     d



a=c()
b=c()
c=c()
d=c()
pvals = fastFET(a,b,c,d)
}
