% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections.R
\name{getGenes}
\alias{getGenes}
\title{Unique Genes in a Library}
\usage{
getGenes(x)
}
\arguments{
\item{x}{A library of sets in List, LongDF or Mat format.}
}
\value{
A character vector of all unique gene names in a library.
}
\description{
Gets unique genes in a library
}
\examples{
#load example gene set libraries
data("encode_example")
unique_encode_genes = getGenes(encode_example)
}
