% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_gene_names.R
\name{HGNCapproved}
\alias{HGNCapproved}
\title{Determine whether there are weights associated with genes.}
\usage{
HGNCapproved(genes, untranslatable.na = F, weighted = F)
}
\arguments{
\item{genes}{a character vector of genes}

\item{untranslatable.na}{boolean, default is false. If a gene does not have an HGNC
mapping, then the function will return an NA for that gene}

\item{weighted}{boolean representing whether genes have weights}
}
\value{
a character vector with genes mapped to approved HGNC symbols
}
\description{
Determine whether there are weights associated with genes.
}
\examples{
geneset = c("FOXO1", "MYC", "GATA3", "BP1")
mapped_geneset = HGNCapproved(geneset, untranslatable.na = T)

}
