% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_formats.R
\name{isWeightedLib}
\alias{isWeightedLib}
\title{Determine whether there are weights associated with genes.}
\usage{
isWeightedLib(x)
}
\arguments{
\item{x}{Genesetr library object.}
}
\value{
boolean
}
\description{
Determine whether there are weights associated with genes.
}
\examples{
gmt = loadGMT('/user/libs/mylib.tsv')
print(isWeightedLib)
}
