% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_gene_names.R
\name{lib2HGNC}
\alias{lib2HGNC}
\title{Map all genes in a genesetr library object to HGNC-approved symbols}
\usage{
lib2HGNC(lib, untranslatable.na = F)
}
\arguments{
\item{lib}{a genesetr library object}

\item{untranslatable.na}{boolean, default is false. If a gene does not have an HGNC
mapping, then the function will return an NA for that gene}
}
\value{
a genesetr library object with all genes mapped to approved HGNC symbols
}
\description{
Map all genes in a genesetr library object to HGNC-approved symbols
}
\examples{
lib = genesetr::chea_example
mapped_lib = lib2HGNC(lib, untranslatable.na = T)

}
