% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_libs.R
\name{loadGMT}
\alias{loadGMT}
\title{Load GMT from disk.}
\usage{
loadGMT(filename)
}
\arguments{
\item{filename}{The path to the file.}
}
\value{
A named list of character vectors.

A GMT is a tab-separated file where each line can have a variable number of terms where
each line corresponds to a set (e.g. a gene set). The first term in each line
corresponds to the set label.
}
\description{
Load GMT from disk.
}
\examples{
gmt = loadGMT("user/dir/my.gmt")
}
