% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_libs.R
\name{loadLongDF}
\alias{loadLongDF}
\title{Load a LongDF format gene set library from disk.}
\usage{
loadLongDF(filename)
}
\arguments{
\item{filename}{The path to the file.}
}
\value{
A two-column data frame representing a gene set library.

The first column is the gene members and the second caolumn is the gene set names.
}
\description{
Load a LongDF format gene set library from disk.
}
\examples{
gmt = loadLongDF("user/dir/my.gmt")
}
