% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_libs.R
\name{loadMat}
\alias{loadMat}
\title{Load a MatLib from disk.}
\usage{
loadMat(filename)
}
\arguments{
\item{filename}{The path to the file.}
}
\value{
A sparse matrix representing the gene library.

A MatLib is a (usually sparse) matrix representing a gene set library. The column labels are gene sets and the rows are genes. Membership of a gene to a gene set is designated by a non-zero entry in the matrix.
}
\description{
Load a MatLib from disk.
}
\examples{
gmt = loadMat("user/dir/my.gmt")
}
