% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{oddsRatio}
\alias{oddsRatio}
\title{Odds Ratio for Multiple Contigency Tables}
\usage{
oddsRatio(a, b, c, d)
}
\arguments{
\item{a}{A vector of values where \eqn{a_i} corresponds to \eqn{t_i}}

\item{b}{A vector of values where b_{i} corresponds to t_{i}}

\item{c}{A vector of values where c_{i} corresponds to t_{i}}

\item{d}{A vector of values where d_{i} corresponds to t_{i}}
}
\value{
A vector of odds ratios R={r_{1}, r_{2},...,r_{n-1},r_{n}} where r_i corresponds to the odds ratio of t_i.
}
\description{
Computes a vector of odds ratios given a set of contingency tables.
}
\details{
Given a set of contingency tables T, where
T = \{t_{1},t_{2},...,t_{n-1},t_{n}\}, with each t_i of the form:

\tabular{cccccc}{
 \tab setA \tab\tab ¬setA \tab \tab \cr
 setB \tab a \tab\tab b \tab | \tab a + b\cr
 ¬setB \tab c \tab\tab d \tab | \tab c + d\cr
  \tab  a + c  \tab\tab  b + d  \tab | \tab a + b + c + d\cr
  }

return a vector of odds ratios R={r_{1}, r_{2},...,r_{n-1},r_{n}} where r_i corresponds to the odds ratio of t_i.
}
\examples{

Given three contingency tables:
     setA  ¬setA          setA  ¬setA         setB  ¬setB
setB  3     100        setC  5     6        setD  20     45
¬setB  123   500       ¬setC  10    100     ¬setD  60     1000



a=c(3, 5, 20)
b=c(100, 6, 45)
c=c(123, 10, 60)
d=c(500, 100, 1000)
or = oddsRatio(a,b,c,d)
}
