% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections.R
\name{pairwiseContingTables}
\alias{pairwiseContingTables}
\title{Contingency Tables}
\usage{
pairwiseContingTables(lib1, lib2, background = NULL)
}
\arguments{
\item{lib1}{A library of gene sets in List, LongDF or Mat format.}

\item{lib2}{A library of gene sets in List, LongDF or Mat format.}

\item{background}{The size of the universe of genes under consideration. User may set to a constant. Default is NULL which results in background being set to the size of the union of the unique genes of lib1 and lib2.}
}
\value{
A data frame with the following fields:
\describe{
  \item{a}{x}
  \item{b}{x}
  \item{c}{x}
  \item{d}{x}
  \item{background}{x}
  \item{len_set1}{x}
  \item{len_set2}{x}
  ...
}
}
\description{
2x2 contingency tables
}
\examples{
#load example gene set libraries
data("encode_example","chea_example")
conting_tab = pairwiseContingTables(lib1, lib2,background= 30000)
conting_tab = pairwiseContingTables(lib1, lib2)
}
