% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections.R
\name{pairwiseIntersects}
\alias{pairwiseIntersects}
\title{All Pairwise Intersections Between Two Libraries
Quickly computes the size of the intersection between all pairs of sets within a library.}
\usage{
pairwiseIntersects(lib1, lib2)
}
\arguments{
\item{lib1}{A library of gene sets in List, LongDF or Mat format.}

\item{lib2}{A library of gene sets in List, LongDF or Mat format.}
}
\value{
A matrix with lib1 gene set names along the rows and lib2 gene set names on the columns populated with the sizes of the intersections.
}
\description{
All Pairwise Intersections Between Two Libraries
Quickly computes the size of the intersection between all pairs of sets within a library.
}
\examples{
#load example gene set libraries
data("encode_example", "chea_example")
intersects = pairwiseIntersects(encode_example, chea_example)
}
