% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pairwiseSetOverlap}
\alias{pairwiseSetOverlap}
\title{All Pairwise Comparisons Between Gene Sets of Two Libraries}
\usage{
pairwiseSetOverlap(lib1, lib2, background = NULL, method = "FET")
}
\arguments{
\item{lib1}{A library of gene sets in List, LongDF or Mat format.}

\item{lib2}{A library of gene sets in List, LongDF or Mat format.}

\item{background}{The size of the "universe" of genes to use in building contingency tables from pairwise intersections of gene sets in the two libraries. Default is NULL-- background will be computed as the size of the union of unique genes in lib1 and lib2.}

\item{method}{Metric by which to assess gene set association. Options are "FET" and "OR".}
}
\value{
A data frame with all pairs of genesets and association metrics
}
\description{
All Pairwise Comparisons Between Gene Sets of Two Libraries
}
\examples{
#load example gene set libraries
data("encode_example", "chea_example")
intersects = pairwiseSetOverlap(encode_example, chea_example)
}
