% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_libs.R
\name{removeDupes}
\alias{removeDupes}
\title{Remove duplicate genes from gene set.}
\usage{
removeDupes(x)
}
\arguments{
\item{filename}{Genesetr library object.}
}
\value{
Returns library object with duplicate genes removed.
}
\description{
Remove duplicate genes from gene set.
}
\examples{
gmt = loadLongDF('/user/libs/mylib.tsv')
myLib = removeDupes(myLib)
}
