% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_libs.R
\name{removeEmptySets}
\alias{removeEmptySets}
\title{Remove empty gene sets from a library}
\usage{
removeEmptySets(x)
}
\arguments{
\item{filename}{Genesetr library object.}
}
\value{
Returns library object with empty sets removed.
}
\description{
Remove empty gene sets from a library
}
\examples{
gmt = loadLongDF('/user/libs/mylib.tsv')
myLib = removeEmptySets(myLib)
}
