% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_formats.R
\name{removeLibWeights}
\alias{removeLibWeights}
\title{Remove weights associated with gene set library.}
\usage{
removeLibWeights(x)
}
\arguments{
\item{x}{Genesetr library object.}
}
\value{
library with weights removed

Some genesetr functions do not accomodated weighted gene set libraries.
}
\description{
Remove weights associated with gene set library.
}
\examples{
gmt = loadGMT('/user/libs/mylib.tsv')
print(isWeightedLib)
}
