% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_formats.R
\name{removeSetWeights}
\alias{removeSetWeights}
\title{Remove weights associated with a gene set.}
\usage{
removeSetWeights(x)
}
\arguments{
\item{x}{A character vector representing a gene set}
}
\value{
Character vector with weights removed

In genesetr, weights are represented by genename,weight
e.g. FOXO1,0.564232
}
\description{
Remove weights associated with a gene set.
}
\examples{
geneset = c("MYC,2.4", "FOXO1,5.2341", "BP1,0.43256")
unweighted_geneset = removeSetWeights(geneset)
}
